Set up http to websocket connection
