+++ ch4/src/pages/index.html
@@ -1,0 +1,73 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Websocket Test</title>
+</head>
+
+<body>
+    <h1>Message Test</h1>
+    <pre id="messages" style="height: 400px; overflow: scroll"></pre>
+    <input type="text" id="messageBox" placeholder="Type your message here"
+        style="display: block; width: 100%; margin-bottom: 10px; padding: 10px;" />
+    <button id="send" title="'Send Message!" style="width: 100%; height: 30px;">Send Message</button>
+
+
+    <script>
+        (function () {
+            const sendBtn = document.querySelector('#send');
+            const messages = document.querySelector('#messages');
+            const messageBox = document.querySelector('#messageBox');
+
+            let ws;
+
+            function showMessage(data) {
+                if (typeof data === 'string') {
+                    messages.textContent += `\n\n${data}`;
+                } else if (data instanceof Blob) {
+                    const reader = new FileReader();
+                    reader.onload = () => {
+                        messages.textContent += `\n\n${reader.result}`;
+                    };
+                    reader.readAsText(data);
+                }
+                messages.scrollTop = messages.scrollHeight;
+                messageBox.value = '';
+            }
+
+            function init() {
+                if (ws) {
+                    ws.onerror = ws.onopen = ws.onclose = null;
+                    ws.close();
+                }
+
+                ws = new WebSocket('ws://localhost:3333');
+                ws.binaryType = 'text';
+                ws.onopen = () => {
+                    console.log('Connection opended!');
+                }
+                ws.onmessage = ({ data }) => showMessage(data);
+                ws.onclose = function () {
+                    ws = null;
+                }
+            }
+
+            sendBtn.onclick = function () {
+                if (!ws) {
+                    showMessage('No webSocket Connection');
+                    return;
+                }
+
+                ws.send(messageBox.value);
+                showMessage(messageBox.value);
+            }
+
+            init();
+        })();
+    </script>
+</body>
+
+</html>
